# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/children/{childId}/days/{date}": {
      "get": {
        "operationId": "getDayPlan",
        "parameters": [
          {
            "name": "childId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayPlanDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Thực đơn một ngày, tự lập nếu chưa có (UC-04)",
        "tags": [
          "meal-planning"
        ]
      }
    },
    "/api/v1/meals/{mealId}": {
      "patch": {
        "operationId": "updateMeal",
        "parameters": [
          {
            "name": "mealId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMealDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Đánh dấu đã chuẩn bị (FR-025)",
        "tags": [
          "meal-planning"
        ]
      }
    },
    "/api/v1/meals/{mealId}/swap-suggestions": {
      "get": {
        "operationId": "getSwapSuggestions",
        "parameters": [
          {
            "name": "mealId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reason",
            "required": false,
            "in": "query",
            "schema": {
              "default": "missing_ingredient",
              "type": "string",
              "enum": [
                "missing_ingredient",
                "disliked",
                "faster",
                "other"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwapSuggestionsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Gợi ý đổi món cho một bữa (UC-06)",
        "tags": [
          "meal-planning"
        ]
      }
    },
    "/api/v1/meals/{mealId}/swap": {
      "post": {
        "operationId": "swapMeal",
        "parameters": [
          {
            "name": "mealId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwapDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Đổi món cho bữa, kiểm tra lại bộ lọc an toàn (FR-045)",
        "tags": [
          "meal-planning"
        ]
      }
    },
    "/api/v1/children/{childId}/dishes": {
      "get": {
        "operationId": "listDishes",
        "parameters": [
          {
            "name": "childId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Tên món hoặc nguyên liệu, không cần dấu",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "chip",
            "required": false,
            "in": "query",
            "schema": {
              "default": "all",
              "type": "string",
              "enum": [
                "all",
                "chicken",
                "fish",
                "beef",
                "pork",
                "legume",
                "snack",
                "custom"
              ]
            }
          },
          {
            "name": "fresh",
            "required": false,
            "in": "query",
            "description": "Chỉ món chưa ăn trong 7 ngày",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Thư viện món đã lọc cho bé (UC-07)",
        "tags": [
          "meal-planning"
        ]
      }
    },
    "/api/v1/children/{childId}/dishes/{dishId}": {
      "get": {
        "operationId": "getRecipe",
        "parameters": [
          {
            "name": "childId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dishId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 4,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Công thức món theo bé và giai đoạn (UC-05)",
        "tags": [
          "meal-planning"
        ]
      }
    },
    "/api/v1/children/{childId}/custom-dishes": {
      "post": {
        "operationId": "createCustomDish",
        "parameters": [
          {
            "name": "childId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomDishInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo món của bạn (UC-23)",
        "tags": [
          "meal-planning"
        ]
      }
    },
    "/api/v1/children/{childId}/custom-dishes/{dishId}": {
      "get": {
        "operationId": "getCustomDish",
        "parameters": [
          {
            "name": "childId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dishId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomDishFormDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Món của bạn, để sửa (FR-135)",
        "tags": [
          "meal-planning"
        ]
      },
      "put": {
        "operationId": "updateCustomDish",
        "parameters": [
          {
            "name": "childId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dishId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomDishInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sửa món của bạn (FR-135)",
        "tags": [
          "meal-planning"
        ]
      },
      "delete": {
        "operationId": "deleteCustomDish",
        "parameters": [
          {
            "name": "childId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dishId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa món của bạn (FR-136)",
        "tags": [
          "meal-planning"
        ]
      }
    },
    "/api/v1/children/{childId}/weeks/{weekStart}": {
      "get": {
        "operationId": "getWeekPlan",
        "parameters": [
          {
            "name": "childId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "weekStart",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeekPlanDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Thực đơn tuần (Thứ Hai → Chủ nhật) và chỉ số (UC-12)",
        "tags": [
          "meal-planning"
        ]
      }
    },
    "/api/v1/children/{childId}/weeks/{weekStart}/generate": {
      "post": {
        "operationId": "generateWeekPlan",
        "parameters": [
          {
            "name": "childId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "weekStart",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateWeekDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeekPlanDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lên thực đơn tuần (UC-13)",
        "tags": [
          "meal-planning"
        ]
      }
    },
...
  },
  "info": {
    "title": "App thực đơn ăn dặm API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "MealDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slot": {
            "type": "string",
            "enum": [
              "breakfast",
              "morning_snack",
              "lunch",
              "afternoon_snack",
              "dinner",
              "extra_snack"
            ]
          },
          "time": {
            "type": "string",
            "example": "11:00"
          },
          "status": {
            "type": "string",
            "enum": [
              "planned",
              "prepared",
              "eaten",
              "refused",
              "skipped"
            ]
          },
          "texture": {
            "type": "string",
            "enum": [
              "puree_smooth",
              "mashed",
              "lumpy",
              "minced_soft",
              "family"
            ]
          },
          "portionText": {
            "type": "string"
          },
          "dish": {
            "$ref": "#/components/schemas/MealDishDto"
          },
          "newIngredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamedIngredientDto"
            }
          },
          "loggedBy": {
            "nullable": true,
            "description": "Ai đã ghi nhận (FR-118)",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LoggedByDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "slot",
          "time",
          "status",
          "texture",
          "portionText",
          "dish",
          "newIngredients",
          "loggedBy"
        ]
      },
      "SlotDto": {
        "type": "object",
        "properties": {
          "slot": {
            "type": "string",
            "enum": [
              "breakfast",
              "morning_snack",
              "lunch",
              "afternoon_snack",
              "dinner",
              "extra_snack"
            ]
          },
          "time": {
            "type": "string"
          }
        },
        "required": [
          "slot",
          "time"
        ]
      },
      "DayPlanDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "plannable": {
            "type": "boolean"
          },
          "meals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MealDto"
            }
          },
          "unfilledSlots": {
            "description": "Bữa chưa tìm được món an toàn",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SlotDto"
            }
          },
          "nextMealId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "date",
          "plannable",
          "meals",
          "unfilledSlots",
          "nextMealId"
        ]
      },
      "UpdateMealDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "prepared"
            ],
            "description": "Chỉ cho phép đánh dấu đã chuẩn bị (FR-025)"
          }
        },
        "required": [
          "status"
        ]
      },
      "MealStatusDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "planned",
              "prepared",
              "eaten",
              "refused",
              "skipped"
            ]
          }
        },
        "required": [
          "id",
          "status"
        ]
      },
...
      "SwapMealDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slot": {
            "type": "string",
            "enum": [
              "breakfast",
              "morning_snack",
              "lunch",
              "afternoon_snack",
              "dinner",
              "extra_snack"
            ]
          },
          "time": {
            "type": "string"
          },
          "dish": {
            "$ref": "#/components/schemas/SwapMealDishDto"
          }
        },
        "required": [
          "id",
          "slot",
          "time",
          "dish"
        ]
      },
...
      "DayProteinDto": {
        "type": "object",
        "properties": {
          "slot": {
            "type": "string",
            "enum": [
              "breakfast",
              "morning_snack",
              "lunch",
              "afternoon_snack",
              "dinner",
              "extra_snack"
            ]
          },
          "protein": {
            "type": "string",
            "enum": [
              "fish",
              "chicken",
              "beef",
              "pork",
              "legume",
              "egg"
            ]
          }
        },
        "required": [
          "slot",
          "protein"
        ]
      },
...
      "ExclusionSummaryDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "byReason": {
            "$ref": "#/components/schemas/ExclusionCountsDto"
          }
        },
        "required": [
          "total",
          "byReason"
        ]
      },
      "SwapSuggestionsDto": {
        "type": "object",
        "properties": {
          "meal": {
            "$ref": "#/components/schemas/SwapMealDto"
          },
          "reason": {
            "type": "string",
            "enum": [
              "missing_ingredient",
              "disliked",
              "faster",
              "other"
            ]
          },
          "ranked": {
            "description": "Món đầu tiên là “Phù hợp nhất”",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SwapCandidateDto"
            }
          },
          "otherMains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayProteinDto"
            }
          },
          "excluded": {
            "$ref": "#/components/schemas/ExclusionSummaryDto"
          },
          "relaxedWindowDays": {
            "type": "number",
            "nullable": true,
            "enum": [
              3
            ]
          }
        },
        "required": [
          "meal",
          "reason",
          "ranked",
          "otherMains",
          "excluded",
          "relaxedWindowDays"
        ]
      },
      "SwapDto": {
        "type": "object",
        "properties": {
          "dishId": {
            "type": "string",
            "example": "dish_chao_bo_cai_bo_xoi"
          },
          "reason": {
            "type": "string",
            "enum": [
              "missing_ingredient",
              "disliked",
              "faster",
              "other"
            ]
          },
          "expectedDishId": {
            "type": "string",
            "description": "Món bữa đang có khi phụ huynh mở gợi ý; khác món hiện tại → 409 MEAL_CHANGED (BR-77)"
          }
        },
        "required": [
          "dishId",
          "reason"
        ]
      },
...
      "LibraryDto": {
        "type": "object",
        "properties": {
          "dishes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibraryDishDto"
            }
          },
          "hidden": {
            "description": "Món khớp bộ lọc nhưng không an toàn cho bé",
            "allOf": [
              {
                "$ref": "#/components/schemas/HiddenDishesDto"
              }
            ]
          }
        },
        "required": [
          "dishes",
          "hidden"
        ]
      },
...
      "RecipeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "mealType": {
            "type": "string",
            "enum": [
              "main",
              "snack"
            ]
          },
          "prepMin": {
            "type": "number"
          },
          "cookMin": {
            "type": "number"
          },
          "tool": {
            "type": "string"
          },
          "contentVersion": {
            "type": "number"
          },
          "reviewedBy": {
            "type": "string",
            "nullable": true
          },
          "custom": {
            "type": "boolean",
            "description": "Món của bạn — chưa qua chuyên gia duyệt (BR-85)"
          },
          "stages": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "selectedStage": {
            "type": "number"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeVariantDto"
            }
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeIngredientDto"
            }
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "safetyNotes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allergens": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "egg",
                "cow_milk",
                "peanut",
                "shellfish",
                "fish",
                "wheat",
                "soy",
                "sesame",
                "tree_nut"
              ]
            }
          },
          "foodGroups": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "carb",
                "protein",
                "fat",
                "veg"
              ]
            }
          },
          "exclusion": {
            "type": "string",
            "enum": [
              "allergen",
              "avoid",
              "paused",
              "age",
              "refused",
              "sick_new"
            ],
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "imageUrl",
          "mealType",
          "prepMin",
          "cookMin",
          "tool",
          "contentVersion",
          "reviewedBy",
          "custom",
          "stages",
          "selectedStage",
          "variants",
          "ingredients",
          "steps",
          "safetyNotes",
          "allergens",
          "foodGroups",
          "exclusion"
        ]
      },
...
      "CustomDishInputDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 60
          },
          "mealType": {
            "type": "string",
            "enum": [
              "main",
              "snack"
            ]
          },
          "ingredients": {
            "description": "1–15 nguyên liệu trong danh mục",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomDishIngredientDto"
            }
          },
          "prepMin": {
            "type": "number",
            "minimum": 0,
            "maximum": 180
          },
          "cookMin": {
            "type": "number",
            "minimum": 0,
            "maximum": 240
          },
          "steps": {
            "description": "Tối đa 15 bước, mỗi bước ≤ 300 ký tự",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "mealType",
          "ingredients",
          "prepMin",
          "cookMin",
          "steps"
        ]
      },
...
      "CustomDishFormDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "mealType": {
            "type": "string",
            "enum": [
              "main",
              "snack"
            ]
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomDishFormIngredientDto"
            }
          },
          "prepMin": {
            "type": "number"
          },
          "cookMin": {
            "type": "number"
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "name",
          "mealType",
          "ingredients",
          "prepMin",
          "cookMin",
          "steps"
        ]
      },
...
      "WeekPlanDto": {
        "type": "object",
        "properties": {
          "weekStart": {
            "type": "string",
            "format": "date"
          },
          "plannable": {
            "type": "boolean"
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeekDayDto"
            }
          },
          "stats": {
            "$ref": "#/components/schemas/WeekStatsDto"
          }
        },
        "required": [
          "weekStart",
          "plannable",
          "days",
          "stats"
        ]
      },
      "GenerateWeekDto": {
        "type": "object",
        "properties": {
          "overwrite": {
            "type": "boolean",
            "default": false,
            "description": "Thay các bữa chưa nấu, chưa ghi nhận"
          }
        }
      },
...
    }
  }
}